refactor: move logic to `useRoomToolboxActions`
